+++ pyron-plugin-impl/src/test/scala/com/cloudentity/pyron/plugin/impl/transform/TransformJsonBodyTest.scala
@@ -2,6 +2,7 @@ package com.cloudentity.pyron.plugin.impl.transform
 
 import com.cloudentity.pyron.plugin.impl.transform.TransformJsonBody.{applyBodyTransformations, setJsonBody}
 import com.cloudentity.pyron.plugin.util.value._
+import com.cloudentity.pyron.plugin.util.value.{JsonValueIgnoreNullIfDefault => JsonWrap}
 import com.cloudentity.pyron.test.TestRequestResponseCtx
 import io.vertx.core.buffer.Buffer
 import io.vertx.core.json.{JsonArray, JsonObject}
@@ -17,90 +18,102 @@ class TransformJsonBodyTest extends WordSpec with MustMatchers with TestRequestR
     def emptyBody = new JsonObject()
 
     "set string value in empty body" in {
-      val valAtPath = Map(Path("a") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(StringJsonValue("string"))))
       setJsonBody(valAtPath, Nil)(emptyBody) mustBe emptyBody.put("a", "string")
     }
     "set string value deep in empty body" in {
-      val valAtPath = Map(Path("a", "b") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("a", "b") -> JsonWrap(Option(StringJsonValue("string"))))
       setJsonBody(valAtPath, Nil)(emptyBody) mustBe emptyBody.put("a", emptyBody.put("b", "string"))
     }
     "set json object value in empty body" in {
       val obj = new JsonObject().put("x", "y")
-      val valAtPath = Map(Path("a") -> Option(ObjectJsonValue(obj.copy())))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(ObjectJsonValue(obj.copy()))))
       setJsonBody(valAtPath, Nil)(emptyBody) mustBe emptyBody.put("a", obj.copy())
     }
     "set json array value in empty body" in {
       val arr = new JsonArray().add("x")
-      val valAtPath = Map(Path("a") -> Option(ArrayJsonValue(arr.copy())))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(ArrayJsonValue(arr.copy()))))
       setJsonBody(valAtPath, Nil)(emptyBody) mustBe emptyBody.put("a", arr.copy())
     }
     "set null value in empty body" in {
-      val valAtPath = Map(Path("a") -> Option(NullJsonValue))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(NullJsonValue)))
       setJsonBody(valAtPath, Nil)(emptyBody) mustBe emptyBody.put("a", null.asInstanceOf[String])
     }
 
     def shallowBody = new JsonObject().put("x", "value")
 
     "set string value in shallow" in {
-      val valAtPath = Map(Path("a") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(StringJsonValue("string"))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody.put("a", "string")
     }
     "set string value deep in shallow" in {
-      val valAtPath = Map(Path("a", "b") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("a", "b") -> JsonWrap(Option(StringJsonValue("string"))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody.put("a", new JsonObject().put("b", "string"))
     }
     "set json object value in shallow" in {
       val obj = new JsonObject().put("x", "y")
-      val valAtPath = Map(Path("a") -> Option(ObjectJsonValue(obj.copy())))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(ObjectJsonValue(obj.copy()))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody.put("a", obj.copy())
     }
     "set json array value in shallow" in {
       val arr = new JsonArray().add("x")
-      val valAtPath = Map(Path("a") -> Option(ArrayJsonValue(arr.copy())))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(ArrayJsonValue(arr.copy()))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody.put("a", arr.copy())
     }
     "set null value in shallow" in {
-      val valAtPath = Map(Path("a") -> Option(NullJsonValue))
+      val valAtPath = Map(Path("a") -> JsonWrap(Option(NullJsonValue)))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody.put("a", null.asInstanceOf[String])
     }
 
     "overwrite with string value in shallow" in {
-      val valAtPath = Map(Path("x") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(StringJsonValue("string"))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody.put("x", "string")
     }
     "overwrite with json object value in shallow" in {
       val obj = new JsonObject().put("u", "v")
-      val valAtPath = Map(Path("x") -> Option(ObjectJsonValue(obj.copy())))
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(ObjectJsonValue(obj.copy()))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody.put("x", obj.copy())
     }
     "overwrite with json array value in shallow" in {
       val arr = new JsonArray().add("x")
-      val valAtPath = Map(Path("x") -> Option(ArrayJsonValue(arr.copy())))
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(ArrayJsonValue(arr.copy()))))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody.put("x", arr.copy())
     }
     "overwrite with null value in shallow" in {
-      val valAtPath = Map(Path("x") -> Option(NullJsonValue))
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(NullJsonValue)))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody.put("x", null.asInstanceOf[String])
     }
     "overwrite with null in shallow if reference missing" in {
-      val valAtPath = Map(Path("x") -> None)
+      val valAtPath = Map(Path("x") -> JsonWrap(None))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody.put("x", null.asInstanceOf[String])
     }
     "add null in shallow if reference missing" in {
-      val valAtPath = Map(Path("z") -> None)
+      val valAtPath = Map(Path("z") -> JsonWrap(None))
       setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody.put("z", null.asInstanceOf[String])
     }
     "retain null value in shallow even if nullIfAbsent is false" in {
-      val valAtPath = Map(Path("x") -> Option(NullJsonValue))
-      setJsonBody(valAtPath, Nil, false)(shallowBody) mustBe emptyBody.put("x", null.asInstanceOf[String])
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(NullJsonValue)))
+      setJsonBody(valAtPath, Nil, nullIfAbsent = false)(shallowBody) mustBe emptyBody.put("x", null.asInstanceOf[String])
     }
     "remove existing value in shallow if reference missing if nullIfAbsent is false" in {
-      val valAtPath = Map(Path("x") -> None)
-      setJsonBody(valAtPath, Nil, false)(shallowBody) mustBe emptyBody
+      val valAtPath = Map(Path("x") -> JsonWrap(None))
+      setJsonBody(valAtPath, Nil, nullIfAbsent = false)(shallowBody) mustBe emptyBody
     }
     "omit key in shallow if reference missing if nullIfAbsent is false" in {
-      val valAtPath = Map(Path("z") -> None)
-      setJsonBody(valAtPath, Nil, false)(shallowBody) mustBe shallowBody
+      val valAtPath = Map(Path("z") -> JsonWrap(None))
+      setJsonBody(valAtPath, Nil, nullIfAbsent = false)(shallowBody) mustBe shallowBody
+    }
+    "retain null value in shallow even if ignoreNullIfAbsent is true" in {
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(NullJsonValue), ignoreNullIfAbsent = true))
+      setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody.put("x", null.asInstanceOf[String])
+    }
+    "remove existing value in shallow if reference missing if ignoreNullIfAbsent is true" in {
+      val valAtPath = Map(Path("x") -> JsonWrap(None, ignoreNullIfAbsent = true))
+      setJsonBody(valAtPath, Nil)(shallowBody) mustBe emptyBody
+    }
+    "omit key in shallow if reference missing if ignoreNullIfAbsent is true" in {
+      val valAtPath = Map(Path("z") -> JsonWrap(None, ignoreNullIfAbsent = true))
+      setJsonBody(valAtPath, Nil)(shallowBody) mustBe shallowBody
     }
 
     "remove existing value" in {
@@ -109,13 +122,13 @@ class TransformJsonBodyTest extends WordSpec with MustMatchers with TestRequestR
     }
 
     "remove existing value while adding new" in {
-      val valAtPath = Map(Path("y") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("y") -> JsonWrap(Option(StringJsonValue("string"))))
       val removePath = List(Path("x"))
       setJsonBody(valAtPath, removePath)(shallowBody) mustBe emptyBody.put("y", "string")
     }
 
     "remove existing value while replacing" in {
-      val valAtPath = Map(Path("x") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("x") -> JsonWrap(Option(StringJsonValue("string"))))
       val removePath = List(Path("x"))
       setJsonBody(valAtPath, removePath)(shallowBody) mustBe emptyBody
     }
@@ -123,29 +136,29 @@ class TransformJsonBodyTest extends WordSpec with MustMatchers with TestRequestR
     def complexBody = new JsonObject(""" { "x": "value", "y" : { "z": "value" } } """)
 
     "overwrite with string value in complex body" in {
-      val valAtPath = Map(Path("y", "z") -> Option(StringJsonValue("string")))
+      val valAtPath = Map(Path("y", "z") -> JsonWrap(Option(StringJsonValue("string"))))
       setJsonBody(valAtPath, Nil)(complexBody) mustBe
         new JsonObject(""" { "x": "value", "y" : { "z": "string" } } """)
     }
     "overwrite with json object value in complex body" in {
       val obj = new JsonObject().put("u", "v")
-      val valAtPath = Map(Path("y", "z") -> Option(ObjectJsonValue(obj)))
+      val valAtPath = Map(Path("y", "z") -> JsonWrap(Option(ObjectJsonValue(obj))))
       setJsonBody(valAtPath, Nil)(complexBody) mustBe
         new JsonObject(""" { "x": "value", "y" : { "z": { "u" : "v" } } } """)
     }
     "overwrite with json array value in complex body" in {
       val arr = new JsonArray().add("u")
-      val valAtPath = Map(Path("y", "z") -> Option(ArrayJsonValue(arr)))
+      val valAtPath = Map(Path("y", "z") -> JsonWrap(Option(ArrayJsonValue(arr))))
       setJsonBody(valAtPath, Nil)(complexBody) mustBe
         new JsonObject(""" { "x": "value", "y" : { "z": [ "u" ] } } """)
     }
     "overwrite with null value in complex body" in {
-      val valAtPath = Map(Path("y", "z") -> Option(NullJsonValue))
+      val valAtPath = Map(Path("y", "z") -> JsonWrap(Option(NullJsonValue)))
       setJsonBody(valAtPath, Nil)(complexBody) mustBe
         new JsonObject(""" { "x": "value", "y" : { "z": null } } """)
     }
     "overwrite with null in complex deep if reference missing" in {
-      val valAtPath = Map(Path("y", "z") -> None)
+      val valAtPath = Map(Path("y", "z") -> JsonWrap(None))
       setJsonBody(valAtPath, Nil)(complexBody) mustBe
         new JsonObject(""" { "x": "value", "y" : { "z": null } } """)
     }
@@ -216,12 +229,12 @@ class TransformJsonBodyTest extends WordSpec with MustMatchers with TestRequestR
 
   "TransformJsonBody.applyBodyTransformations" should {
     "return empty buffer if dropping body without set ops" in {
-      val bodyOps = ResolvedBodyOps(set = None, remove = None, drop = Some(true), nullIfAbsent = None)
+      val bodyOps = ResolvedBodyOps(set = Map.empty, remove = None, drop = Some(true), nullIfAbsent = None)
       applyBodyTransformations(bodyOps, new JsonObject().put("x", "y")) mustBe Buffer.buffer()
     }
 
     "return empty buffer if dropping body with set ops" in {
-      val bodyOps = ResolvedBodyOps(set = Some(Map(Path("x") -> Some(StringJsonValue("a")))), remove = None, drop = Some(true), nullIfAbsent = None)
+      val bodyOps = ResolvedBodyOps(set = Map(Path("x") -> JsonWrap(Some(StringJsonValue("a")))), remove = None, drop = Some(true), nullIfAbsent = None)
       applyBodyTransformations(bodyOps, new JsonObject().put("x", "y")) mustBe Buffer.buffer()
     }
   }